+++ src/pages/LoadingPage.jsx
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import LoadingSpinner from "../LoadingSpinner";
 
-
-const LoginSuccess = () => {
-    const [isLoading, setIsLoading] = useState(false); // 로딩 상태
+const LoadingPage = () => {
+  const [isLoading, setIsLoading] = useState(false); // 로딩 상태
   const [isLoginInProgress, setIsLoginInProgress] = useState(true); // 로그인 진행 상태
   const [currentTime, setCurrentTime] = useState(''); // 현재 시간 표시
   const navigate = useNavigate();
@@ -73,7 +71,7 @@ const LoginSuccess = () => {
 
         {/* 로딩 상태일 때만 로딩 아이콘 */}
         {isLoading ? (
-          <LoadingSpinner />
+          <div className="text-4xl animate-spin mb-4">🔄</div>
         ) : (
           <div className="text-gray-500 mb-4">준비 중...</div>
         )}
@@ -84,6 +82,6 @@ const LoginSuccess = () => {
       </div>
     </div>
   );
-}
+};
 
-export default LoginSuccess;
+export default LoadingPage;
